+++ Assets/Scripts/Interfaces/ITakeDamage.cs
@@ -1,4 +1,4 @@
 interface ITakeDamage
 {
-    void TakeDamage(int damage);
+    void TakeDamage(float damage);
 }
